requirements.txt
django>=5.2
fontawesomefree==6.6.0
django-filter
python-decouple
